Implemented Simple Trajectory Prediction
Added a simple trajectory prediction generator. Currently uses a ton of
smaller spheres to draw a line. Not the most efficient at all. Should be
drawn as a line to avoid the unnecessary GameObjects overhead.
Also changed the gravity controller to calculate a gravity vector,
instead of applying gravity to a rigidbody itself. Now it uses its own
function to calculate but still adds the gravity to the gravity list
objects (player) in its FixedUpdate.